Added support for the official Russian localization of Silkroad (4game).
For authorization you need to enter your email from 4game account as login and password from this account. At the first login, be sure to get to the mail and enter the PIN from e-mail in a special box. Then the session will be saved and the emulation of the launcher will work.

In addition, updated Russian translation, added small interface improvements for 125% scaling.